requirements.txt
aldryn-boilerplates==0.8.0
aldryn-forms==4.0.1
biopython==1.76
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
cssselect==1.1.0
cssutils==1.0.2
Django==2.1.15
django-absolute==0.3
django-appconf==1.0.3
django-classy-tags==1.0.0
django-cms==3.7.1
django-crispy-forms==1.8.1
django-emailit==0.2.4
django-filer==1.6.0
django-formtools==2.2
django-js-asset==1.2.2
django-menu-generator==1.0.4
django-mptt==0.11.0
django-polymorphic==2.0.3
django-ranged-response==0.2.0
django-sekizai==1.1.0
django-simple-captcha==0.5.12
django-sizefield==1.0.0
django-tablib==3.2
django-treebeard==4.3.1
djangocms-admin-style==1.5.0
djangocms-attributes-field==1.2.0
djangocms-column==1.9.0
djangocms-file==2.4.0
djangocms-googlemap==1.4.0
djangocms-link==2.5.0
djangocms-picture==2.4.0
djangocms-snippet==2.3.0
djangocms-style==2.3.0
djangocms-text-ckeditor==3.8.0
djangocms-video==2.3.0
easy-thumbnails==2.7
et-xmlfile==1.0.1
gunicorn==20.0.4
html5lib==1.0.1
idna==2.8
jdcal==1.4.1
lxml==4.5.0
numpy==1.18.1
openpyxl==2.4.9
Pillow==7.0.0
premailer==3.6.1
psycopg2-binary==2.8.4
pytz==2019.3
requests==2.22.0
six==1.14.0
sqlparse==0.3.0
tablib==1.0.0
Unidecode==1.0.23
urllib3==1.25.8
webencodings==0.5.1
whitenoise==5.0.1
YURL==1.0.0
